+++ source/Google/MobileAds/MobileAds.targets
@@ -3,19 +3,13 @@
 	<PropertyGroup>
 		<_GoogleMobileAdsAssemblyName>Google.MobileAds, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null</_GoogleMobileAdsAssemblyName>
 		<_GoogleMobileAdsId>FAdM-7.66.0</_GoogleMobileAdsId>
-		<_GoogleUserMessagingPlatformId>GUsrMess-1.1.0</_GoogleUserMessagingPlatformId>
 		<_GoogleMobileAdsSDKBaseFolder>$(XamarinBuildDownloadDir)$(_GoogleMobileAdsId)\Frameworks\GoogleMobileAdsFramework-Current\</_GoogleMobileAdsSDKBaseFolder>		
-		<_GoogleUserMessagingPlatformSDKBaseFolder>$(XamarinBuildDownloadDir)$(_GoogleUserMessagingPlatformId)\Frameworks\Release\</_GoogleUserMessagingPlatformSDKBaseFolder>		
 	</PropertyGroup>
 	<ItemGroup Condition="('$(OutputType)'!='Library' OR '$(IsAppExtension)'=='True')">
 		<XamarinBuildDownload Include="$(_GoogleMobileAdsId)">
 			<Url>https://dl.google.com/dl/cpdc/080245260b12ad1b/Google-Mobile-Ads-SDK-7.66.0.tar.gz</Url>
 			<Kind>Tgz</Kind>
 		</XamarinBuildDownload>		
-		<XamarinBuildDownload Include="$(_GoogleUserMessagingPlatformId)">
-			<Url>https://dl.google.com/dl/cpdc/624b551688fd020a/GoogleUserMessagingPlatform-1.1.0.tar.gz</Url>
-			<Kind>Tgz</Kind>
-		</XamarinBuildDownload>
 		<XamarinBuildRestoreResources Include="_FAdMDownloadedItems" />
 	</ItemGroup>
 	<Target Name="_FAdMDownloadedItems">
@@ -28,12 +22,6 @@
 				<LinkerFlags>-ObjC -lz -lsqlite3</LinkerFlags>
 				<Frameworks>AudioToolbox AVFoundation CFNetwork CoreGraphics CoreMedia CoreTelephony CoreVideo MediaPlayer MessageUI MobileCoreServices QuartzCore Security StoreKit SystemConfiguration</Frameworks>
 				<WeakFrameworks>AdSupport JavaScriptCore SafariServices WebKit</WeakFrameworks>
-			</NativeReference>
-			<!-- From https://dl.google.com/dl/cpdc/624b551688fd020a/GoogleUserMessagingPlatform-1.1.0.tar.gz -->
-			<NativeReference Include="$(_GoogleUserMessagingPlatformSDKBaseFolder)UserMessagingPlatform.framework">
-				<Kind>Framework</Kind>
-				<SmartLink>True</SmartLink>
-				<ForceLoad>True</ForceLoad>
 			</NativeReference>
 		</ItemGroup>
 	</Target>
