+++ source/Google/UserMessagingPlatform/UserMessagingPlatform.targets
@@ -1,0 +1,24 @@
+﻿<?xml version="1.0" encoding="UTF-8"?>
+<Project ToolsVersion="4.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
+	<PropertyGroup>
+		<_GoogleUserMessagingPlatformId>GUsrMess-1.1.0</_GoogleUserMessagingPlatformId>
+		<_GoogleUserMessagingPlatformSDKBaseFolder>$(XamarinBuildDownloadDir)$(_GoogleUserMessagingPlatformId)\Frameworks\Release\</_GoogleUserMessagingPlatformSDKBaseFolder>		
+	</PropertyGroup>
+	<ItemGroup Condition="('$(OutputType)'!='Library' OR '$(IsAppExtension)'=='True')">
+		<XamarinBuildDownload Include="$(_GoogleUserMessagingPlatformId)">
+			<Url>https://dl.google.com/dl/cpdc/624b551688fd020a/GoogleUserMessagingPlatform-1.1.0.tar.gz</Url>
+			<Kind>Tgz</Kind>
+		</XamarinBuildDownload>
+		<XamarinBuildRestoreResources Include="_FAdMDownloadedItems" />
+	</ItemGroup>
+	<Target Name="_FAdMDownloadedItems">
+		<ItemGroup>
+			<!-- From https://dl.google.com/dl/cpdc/624b551688fd020a/GoogleUserMessagingPlatform-1.1.0.tar.gz -->
+			<NativeReference Include="$(_GoogleUserMessagingPlatformSDKBaseFolder)UserMessagingPlatform.framework">
+				<Kind>Framework</Kind>
+				<SmartLink>True</SmartLink>
+				<ForceLoad>True</ForceLoad>
+			</NativeReference>
+		</ItemGroup>
+	</Target>
+</Project>
